ordens dao layer and api added

+++ backend/src/main/java/com/bashkir777/api/data/repositories/SpaceMarineRepository.java
@@ -3,5 +3,5 @@ package com.bashkir777.api.data.repositories;
 import com.bashkir777.api.data.entities.SpaceMarine;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpaceMarineRepository extends JpaRepository<SpaceMarine, Integer> {
+public interface SpaceMarineRepository extends JpaRepository<SpaceMarine, Long> {
 }
